Order NOTPASSED outcomes based on their reason; misc cleanup
Changing the outcome comparison function also changes the way that
solutions get partitioned; e.g., a solution is now only equal to
another solution when they have the same outcomes on the same
tests. If one solution fails an outcome and another errors on it, the
two are no longer equal.

The string comparisons make this noticeably slower than when there
were no comparisons but it's still acceptable; Maybe if the comparison
was over datatypes instead of strings it would be quicker.