Print out the shortest path from each node to the root
One problem: the graph this operates on on has more edges than the
tred'd fault graphs we've been looking at (which is why we use tred in
the first place). Specifically it has an edge between two nodes if one
of the nodes fails strictly less than the other, and every non-root
node fails strictly less tests than the root, which means every one
has a shortest path of 1.