Label decision tree edges w/outcomes; Add edge attributes to Dot
Given the current Dot structure, the edge between a decision tree's
parent and child node can be created in a couple ways: if the child
creates it, it needs to know how to refer to its parent; if the parent
creates it, it needs to know how to refer to its child. And because
the information we want to put on the edge is right by the parent node
it's convenient to let the parent control creating the edge. But it's
a little tricky to thread the child's name back to the parent because
a Branch constructor contains multiple subtrees, so instead the parent
passes a continuation to the child, which the child gives its name
to.

I'm not attached to this API for adding attributes, and I'm definitely
not attached to this way of handling the decision tree's edges.